MODIFY add table images to the project

+++ admin/liste-avis.php
@@ -39,16 +39,14 @@ $totalPages = ceil($totalreviews / _ADMIN_ITEM_PER_PAGE_);
           </tr>
           <tbody>
             <?php foreach ($reviews as $review) { ?>
-            <tr>
-              <td><?= $review["client"]; ?></td>
-              <td><?= $review["comment"]; ?></td>
-              <td><?= $review["note"]; ?></td>
-              <td><a href="modifier-avis.php?id=<?= $review['id'] ?>"><i class="fa-solid fa-pencil"></i></a>
-                | <a href="delete-review.php?id=<?= $review['id'] ?>"
-                  onclick="return confirm('Êtes-vous sûr de vouloir supprimer cet avis ?')"><i
-                    class="fa-solid fa-trash-can"></i></a>
-              </td>
-            </tr>
+              <tr>
+                <td><?= $review["client"]; ?></td>
+                <td><?= $review["comment"]; ?></td>
+                <td><?= $review["note"]; ?></td>
+                <td><a href="modifier-avis.php?id=<?= $review['id'] ?>"><i class="fa-solid fa-pencil"></i></a>
+                  | <a href="delete-review.php?id=<?= $review['id'] ?>" onclick="return confirm('Êtes-vous sûr de vouloir supprimer cet avis ?')"><i class="fa-solid fa-trash-can"></i></a>
+                </td>
+              </tr>
             <?php } ?>
           </tbody>
         </table>
@@ -56,13 +54,15 @@ $totalPages = ceil($totalreviews / _ADMIN_ITEM_PER_PAGE_);
       <nav aria-label="Page navigation reviews">
         <ul class="pagination">
           <?php if ($totalPages > 1) { ?>
-          <?php for ($i = 1; $i <= $totalPages; $i++) { ?>
-          <li class="page-item">
-            <a class="page-link <?php if ($i == $page) { echo " active";} ?>" href="?page=<?php echo $i; ?>">
-              <?php echo $i; ?>
-            </a>
-          </li>
-          <?php } ?>
+            <?php for ($i = 1; $i <= $totalPages; $i++) { ?>
+              <li class="page-item">
+                <a class="page-link <?php if ($i == $page) {
+                                      echo " active";
+                                    } ?>" href="?page=<?php echo $i; ?>">
+                  <?php echo $i; ?>
+                </a>
+              </li>
+            <?php } ?>
           <?php } ?>
         </ul>
       </nav>
